Fixes the issues
Now things work as designed.

In fact, it fixes an annoyance by changing the UX of the camera list: clicking a camera in the camera list no longer let’s you change the size of the custom render. This way, you will no longer edit the rendering size of the wrong camera.